pyproject.toml
[project]
name = "arena-arcade"
version = "1.0.0"
description = "A professional arcade suite built with Pygame"
authors = [
    { name = "Developer", email = "developer@example.com" }
]
dependencies = [
    "pygame-ce>=2.5.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
